+++ test/unit/storage/keyvalue/WrappedIndexedStorage.test.ts
@@ -59,6 +59,10 @@ describe('A WrappedIndexedStorage', (): void => {
         .rejects.toThrow(InternalServerError);
     });
 
+    it('errors when defining types with array references.', async(): Promise<void> => {
+      await expect(storage.defineType('root', { ref: 'id:Type1[]' } as any)).rejects.toThrow(InternalServerError);
+    });
+
     it('errors when defining types with optional references.', async(): Promise<void> => {
       await expect(storage.defineType('root', { ref: 'id:Type1?' } as any)).rejects.toThrow(InternalServerError);
     });
@@ -356,6 +360,32 @@ describe('A WrappedIndexedStorage', (): void => {
       // Corrupt the index. Will break if we change how index keys get generated.
       indexMap.delete(`child/name/${child.name}`);
       await expect(storage.delete('child', child.id)).resolves.toBeUndefined();
+    });
+
+    it('can handle missing virtual keys.', async(): Promise<void> => {
+      // Remove child -> grandchild virtual key
+      delete valueMap.get(root.id)['**child**'][child2.id]['**grandchild**'];
+      await expect(storage.delete('child', child2.id)).resolves.toBeUndefined();
+
+      // Remove root -> child virtual key
+      delete valueMap.get(root.id)['**child**'];
+      await expect(storage.create('grandchild', { bool: false, parent: child.id, notIndexed: 1 }))
+        .rejects.toThrow('Could not find child ');
+      await storage.find('child', { parent: root.id });
+      const newChild = await storage.create('child', { name: 'name', parent: root.id, notIndexed: 1 });
+      await expect(storage.delete('child', child.id))
+        .rejects.toThrow('Could not find child ');
+      await expect(storage.has('child', child.id)).resolves.toBe(false);
+      await expect(storage.get('child', child.id)).resolves.toBeUndefined();
+      await expect(storage.setField('child', child.id, 'name', 'name'))
+        .rejects.toThrow('Could not find child ');
+      const entries: string[] = [];
+      for await (const entry of storage.entries('child')) {
+        entries.push(entry.id);
+      }
+      expect(entries).toHaveLength(2);
+      expect(entries).toContain(newChild.id);
+      expect(entries).toContain(child3.id);
     });
   });
 });